refactor(lang): split prettify.ts into focused modules
No behavior change. Preparation for the spacing-rule deduplication and
parser-helper extraction that follow. Splits the 1660-line file into:

  prettify/out.ts            -- Out buffer
  prettify/tokens.ts         -- classification sets, layout config, utilities
  prettify/rules.ts          -- section-statement dispatch table, kind labels
  prettify/error-listener.ts -- CollectingErrorListener
  prettify/types.ts          -- PrettifyError, PrettifyResult
  prettify/index.ts          -- Formatter class + prettify() entry point

Public surface unchanged: @malloydata/malloy/internal still re-exports
prettify, PrettifyError, PrettifyResult.

Also removes a dead local (savedPrevLine in formatToString) that the linter
flagged after the move; the variable was captured but never read.
